# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountSnapshotsResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AccountSnapshotsResponse",
        "type": "object"
      },
...
      "DecisionRecord": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "$ref": "#/components/schemas/ResearchAuthor"
          },
          "kind": {
            "const": "decision",
            "title": "Kind",
            "type": "string"
          },
          "mode": {
            "enum": [
              "prospective",
              "retrospective",
              "synthetic"
            ],
            "title": "Mode",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/DecisionPayload"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "mode",
          "author",
          "recorded_at",
          "kind",
          "payload"
        ],
        "title": "DecisionRecord",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "InvestmentContext": {
        "additionalProperties": false,
        "properties": {
          "account": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectedAccount"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_decision_ids": {
            "items": {
              "pattern": "^[0-9a-f]{64}$",
              "type": "string"
            },
            "title": "Active Decision Ids",
            "type": "array"
          },
          "context_kind": {
            "const": "current_research_context",
            "title": "Context Kind",
            "type": "string"
          },
          "eligible_record_count": {
            "title": "Eligible Record Count",
            "type": "integer"
          },
          "exported_record_count": {
            "title": "Exported Record Count",
            "type": "integer"
          },
          "future_record_count": {
            "title": "Future Record Count",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "historical_reproducibility": {
            "const": false,
            "title": "Historical Reproducibility",
            "type": "boolean"
          },
          "omitted_record_ids": {
            "items": {
              "pattern": "^[0-9a-f]{64}$",
              "type": "string"
            },
            "title": "Omitted Record Ids",
            "type": "array"
          },
          "omitted_references": {
            "items": {
              "$ref": "#/components/schemas/OmittedReference"
            },
            "title": "Omitted References",
            "type": "array"
          },
          "orders_enabled": {
            "const": false,
            "title": "Orders Enabled",
            "type": "boolean"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/ResearchResponse"
            },
            "title": "Records",
            "type": "array"
          },
          "retired_decision_ids": {
            "items": {
              "pattern": "^[0-9a-f]{64}$",
              "type": "string"
            },
            "title": "Retired Decision Ids",
            "type": "array"
          },
          "review_queue": {
            "items": {
              "$ref": "#/components/schemas/ReviewQueueItem"
            },
            "title": "Review Queue",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "sizing_validated": {
            "const": false,
            "title": "Sizing Validated",
            "type": "boolean"
          },
          "snapshot_freshness": {
            "discriminator": {
              "mapping": {
                "fresh": "#/components/schemas/SelectedFreshness",
                "future": "#/components/schemas/SelectedFreshness",
                "not_selected": "#/components/schemas/UnselectedFreshness",
                "stale": "#/components/schemas/SelectedFreshness"
              },
              "propertyName": "status"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/UnselectedFreshness"
              },
              {
                "$ref": "#/components/schemas/SelectedFreshness"
              }
            ],
            "title": "Snapshot Freshness"
          },
          "truncated_count": {
            "title": "Truncated Count",
            "type": "integer"
          },
          "trust_instructions": {
            "items": {
              "type": "string"
            },
            "title": "Trust Instructions",
            "type": "array"
          },
          "unresolved_questions": {
            "items": {
              "$ref": "#/components/schemas/UnresolvedQuestions"
            },
            "title": "Unresolved Questions",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "context_kind",
          "generated_at",
          "historical_reproducibility",
          "orders_enabled",
          "sizing_validated",
          "trust_instructions",
          "records",
          "eligible_record_count",
          "exported_record_count",
          "truncated_count",
          "omitted_record_ids",
          "future_record_count",
          "omitted_references",
          "account",
          "snapshot_freshness",
          "active_decision_ids",
          "retired_decision_ids",
          "review_queue",
          "unresolved_questions"
        ],
        "title": "InvestmentContext",
        "type": "object"
      },
...
      "OmittedReference": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "enum": [
              "future_record",
              "record_limit"
            ],
            "title": "Reason",
            "type": "string"
          },
          "record_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Record Id",
            "type": "string"
          },
          "reference_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Reference Id",
            "type": "string"
          }
        },
        "required": [
          "record_id",
          "reference_id",
          "reason"
        ],
        "title": "OmittedReference",
        "type": "object"
      },
...
      "ResearchResponse": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Id",
            "type": "string"
          },
          "record": {
            "discriminator": {
              "mapping": {
                "decision": "#/components/schemas/DecisionRecord",
                "evidence": "#/components/schemas/EvidenceRecord",
                "hypothesis": "#/components/schemas/HypothesisRecord",
                "review": "#/components/schemas/ReviewRecord"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/EvidenceRecord"
              },
              {
                "$ref": "#/components/schemas/HypothesisRecord"
              },
              {
                "$ref": "#/components/schemas/DecisionRecord"
              },
              {
                "$ref": "#/components/schemas/ReviewRecord"
              }
            ],
            "title": "Record"
          }
        },
        "required": [
          "id",
          "record"
        ],
        "title": "ResearchResponse",
        "type": "object"
      },
...
      "ReviewQueueItem": {
        "additionalProperties": false,
        "properties": {
          "decision_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Decision Id",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "review_after": {
            "format": "date-time",
            "title": "Review After",
            "type": "string"
          },
          "review_ids": {
            "items": {
              "pattern": "^[0-9a-f]{64}$",
              "type": "string"
            },
            "title": "Review Ids",
            "type": "array"
          },
          "sizing_validated": {
            "const": false,
            "title": "Sizing Validated",
            "type": "boolean"
          },
          "status": {
            "const": "proposed",
            "title": "Status",
            "type": "string"
          },
          "unresolved_questions": {
            "items": {
              "type": "string"
            },
            "title": "Unresolved Questions",
            "type": "array"
          }
        },
        "required": [
          "decision_id",
          "review_after",
          "reasons",
          "unresolved_questions",
          "review_ids",
          "status",
          "sizing_validated"
        ],
        "title": "ReviewQueueItem",
        "type": "object"
      },
...
      "SelectedAccount": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Id",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/AccountSnapshot"
          }
        },
        "required": [
          "id",
          "snapshot"
        ],
        "title": "SelectedAccount",
        "type": "object"
      },
      "SelectedFreshness": {
        "additionalProperties": false,
        "properties": {
          "atomic_account_instant": {
            "const": false,
            "title": "Atomic Account Instant",
            "type": "boolean"
          },
          "collection_span_seconds": {
            "title": "Collection Span Seconds",
            "type": "number"
          },
          "completed_age_seconds": {
            "title": "Completed Age Seconds",
            "type": "number"
          },
          "max_age_seconds": {
            "title": "Max Age Seconds",
            "type": "integer"
          },
          "oldest_observation_age_seconds": {
            "title": "Oldest Observation Age Seconds",
            "type": "number"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "sizing_validated": {
            "const": false,
            "title": "Sizing Validated",
            "type": "boolean"
          },
          "snapshot_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Snapshot Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "fresh",
              "stale",
              "future"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "snapshot_id",
          "completed_age_seconds",
          "oldest_observation_age_seconds",
          "collection_span_seconds",
          "max_age_seconds",
          "atomic_account_instant",
          "sizing_validated",
          "reasons"
        ],
        "title": "SelectedFreshness",
        "type": "object"
      },
      "SnapshotSummary": {
        "additionalProperties": false,
        "properties": {
          "account_seq": {
            "title": "Account Seq",
            "type": "string"
          },
          "account_type": {
            "title": "Account Type",
            "type": "string"
          },
          "collection_completed_at": {
            "format": "date-time",
            "title": "Collection Completed At",
            "type": "string"
          },
          "collection_started_at": {
            "format": "date-time",
            "title": "Collection Started At",
            "type": "string"
          },
          "holding_count": {
            "title": "Holding Count",
            "type": "integer"
          },
          "id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Id",
            "type": "string"
          },
          "open_order_count": {
            "title": "Open Order Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "account_seq",
          "account_type",
          "collection_started_at",
          "collection_completed_at",
          "holding_count",
          "open_order_count"
        ],
        "title": "SnapshotSummary",
        "type": "object"
      },
...
      "UnresolvedQuestions": {
        "additionalProperties": false,
        "properties": {
          "decision_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Decision Id",
            "type": "string"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "decision_id",
          "questions"
        ],
        "title": "UnresolvedQuestions",
        "type": "object"
      },
      "UnselectedFreshness": {
        "additionalProperties": false,
        "properties": {
          "max_age_seconds": {
            "title": "Max Age Seconds",
            "type": "integer"
          },
          "sizing_validated": {
            "const": false,
            "title": "Sizing Validated",
            "type": "boolean"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "null"
          },
          "status": {
            "const": "not_selected",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "snapshot_id",
          "max_age_seconds",
          "sizing_validated"
        ],
        "title": "UnselectedFreshness",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only local records. No model execution or brokerage orders.",
    "title": "Investment workspace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/account-snapshots": {
      "get": {
        "operationId": "list_account_snapshots",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSnapshotsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Snapshots"
      }
    },
    "/api/v1/context": {
      "get": {
        "operationId": "get_context",
        "parameters": [
          {
            "in": "query",
            "name": "snapshot_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9a-f]{64}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshot Id"
            }
          },
          {
            "in": "query",
            "name": "max_records",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Max Records",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentContext"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Context"
      }
    },
...
    "/api/v1/research/{id}": {
      "get": {
        "operationId": "get_research",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Research Record"
      }
    }
  }
}